\name{calculate_neighborhood_histograms}
\alias{calculate_neighborhood_histograms}
\title{Calculated neighborhood stats.}
\usage{
calculate_neighborhood_histograms(posts)
}
\arguments{
  \item{posts}{the instagram posts, including
  neighborhoods}
}
\description{
Calculated neighborhood stats.
}
\examples{
data(posts_sample)
data(posts_sample)
temporary_path <- tempdir()
calculate_neighborhood_histograms(posts_sample)
}

