\name{calculate_neighborhood_statistics}
\alias{calculate_neighborhood_statistics}
\title{Calculate the neighborhood statistics for use in sonification.}
\usage{
calculate_neighborhood_statistics(posts)
}
\arguments{
  \item{posts}{instagram posts with neighborhoods}
}
\description{
Calculate the neighborhood statistics for use in
sonification.
}

