\name{constrain_to_nyc}
\alias{constrain_to_nyc}
\title{Constrains data to the New York City area (i.e. removes bad data).}
\usage{
constrain_to_nyc(dat)
}
\arguments{
  \item{dat}{data containing geographic coordinates}
}
\description{
Constrains data to the New York City area (i.e. removes bad
data).
}

