\name{flatten_nyc_neighborhood_geodata}
\alias{flatten_nyc_neighborhood_geodata}
\title{Flattens NYC neighborhood data.}
\usage{
flatten_nyc_neighborhood_geodata(nyc_neighborhoods_geodata)
}
\arguments{
  \item{nyc_neighborhoods_geodata}{geodata format of NYC
  neighborhoods}
}
\description{
Flattens NYC neighborhood data.
}

