\docType{package}
\name{instasound}
\alias{instasound}
\alias{instasound-package}
\title{Exploratory analysis of instagram data in NYC.}
\description{
Note that there are several early-phase data loading and
cleaning methods. Most of these should not be used
directly. Package data should be used, instead.
}

