\name{load_instagram_data}
\alias{load_instagram_data}
\title{Loads the source instagram data and constrains it to the fields we care about.}
\usage{
load_instagram_data(path)
}
\arguments{
  \item{path}{the path to the instagram data}
}
\description{
Loads the source instagram data and constrains it to the
fields we care about.
}

