\name{melt_neighborhood_stats}
\alias{melt_neighborhood_stats}
\title{Melts neighborhood stats into a form that makes it easier to create some types of plots.}
\usage{
melt_neighborhood_stats(neighborhood_stats)
}
\arguments{
  \item{neighborhood_stats}{generated from posts}
}
\description{
Melts neighborhood stats into a form that makes it easier
to create some types of plots.
}

