\name{plot_instagram_exploration}
\alias{plot_instagram_exploration}
\title{Plots some exploratory analysis of the data.}
\usage{
plot_instagram_exploration(dat, path = ".")
}
\arguments{
  \item{dat}{instagram data, assumed to be the result of
  prepare_instagram_data}
}
\description{
Plots some exploratory analysis of the data.
}

