\name{plot_instagram_map_with_tod}
\alias{plot_instagram_map_with_tod}
\title{Plots geographic isntagram with TOD.}
\usage{
plot_instagram_map_with_tod(posts)
}
\arguments{
  \item{posts}{instagram posts}
}
\description{
Plots geographic isntagram with TOD.
}
\examples{
data(posts_sample)
plot_instagram_map_with_tod(posts_sample)
}

