\name{plot_instagram_point_base}
\alias{plot_instagram_point_base}
\title{Shared function for plotting instagram points. Uses in other plots.}
\usage{
plot_instagram_point_base(posts)
}
\arguments{
  \item{posts}{instagram posts}
}
\description{
Shared function for plotting instagram points. Uses in
other plots.
}

