\name{plot_instagram_points}
\alias{plot_instagram_points}
\title{Plots instagram map with no geographic map underneath.}
\usage{
plot_instagram_points(posts)
}
\arguments{
  \item{posts}{instagram posts}
}
\description{
Plots instagram map with no geographic map underneath.
}
\examples{
data(posts_sample)
plot_instagram_points(posts_sample)
}

