\name{plot_instagram_points_with_average}
\alias{plot_instagram_points_with_average}
\title{Plots an instagram map with no geographic map underneath, and with the average location.}
\usage{
plot_instagram_points_with_average(posts)
}
\arguments{
  \item{posts}{instagram posts}
}
\description{
Plots an instagram map with no geographic map underneath,
and with the average location.
}
\examples{
data(posts_sample)
plot_instagram_points_with_average(posts_sample)
}

