\name{plot_instagram_points_with_tod}
\alias{plot_instagram_points_with_tod}
\title{Plots instagram map with no geographic map underneath. Use color to represent the time of day.}
\usage{
plot_instagram_points_with_tod(posts)
}
\arguments{
  \item{posts}{instagram posts}
}
\description{
Plots instagram map with no geographic map underneath. Use
color to represent the time of day.
}
\examples{
data(posts_sample)
plot_instagram_points_with_tod(posts_sample)
}

