\name{plot_ny_map}
\alias{plot_ny_map}
\title{Creates a plot of New York City, based on the five counties corresponding to the five boroughs.}
\usage{
plot_ny_map(counties = c("kings", "queens", "new york", "bronx", "richmond"))
}
\arguments{
  \item{counties}{which NYC counties to include. Defaults
  to all five borough counties.}
}
\description{
Creates a plot of New York City, based on the five counties
corresponding to the five boroughs.
}
\examples{
plot_ny_map()
}

