\name{scale_color_tod}
\alias{scale_color_tod}
\title{Creates a gradient color scale for time of day.}
\usage{
scale_color_tod(night = "navy", day = "#FCD116")
}
\arguments{
  \item{night}{the color to use for nighttime}

  \item{day}{the color to use for daytime}
}
\description{
Creates a gradient color scale for time of day.
}

