\name{valid_neighborhoods}
\alias{valid_neighborhoods}
\title{Stores the list of valid neighborhoods in a JSON file.}
\usage{
valid_neighborhoods(path = "../data/valid_neighborhoods.json")
}
\arguments{
  \item{path}{in which to store the file}
}
\description{
Stores the list of valid neighborhoods in a JSON file.
}

