\name{write_instagram_data}
\alias{write_instagram_data}
\title{Writes instagram data to a new TSV file.}
\usage{
write_instagram_data(instagram_data, path)
}
\arguments{
  \item{instagram_data}{e.g. loaded using
  load_instagram_data}

  \item{path}{in which to store the data}
}
\description{
Writes instagram data to a new TSV file.
}

