\name{plot_neighborhood_tod_polar}
\alias{plot_neighborhood_tod_polar}
\title{Generates a polar TOD plot for each neighborhood.}
\usage{
plot_neighborhood_tod_polar(posts)
}
\arguments{
  \item{posts}{instagram posts with neighborhoods}
}
\description{
Generates a polar TOD plot for each neighborhood.
}

