\name{plot_ny_map}
\alias{plot_ny_map}
\title{Creates a plot of New York City, based on the five counties corresponding to the five boroughs.}
\usage{
plot_ny_map(counties = c("kings", "queens", "new york", "bronx", "richmond"))
}
\description{
Creates a plot of New York City, based on the five counties
corresponding to the five boroughs.
}

